requirements.txt
geopy
knmi-py
requests
edr-pydantic
fastapi
pandas
